cdk-lakeformation-permissions/source/requirements.txt
aws-cdk-lib>=2.74.0
constructs>=10.0.0
types-boto3